+++ docs/releases.md
@@ -8,12 +8,22 @@ title: Release Notes
 This file contains changes and migration steps for the Graphback project. 
 Please follow individual releases for more information.
 
+# 0.16
+
+### New Features
+
+* Requiring `_id: GraphbackObjectID` primary key for MongoDB ([#1769](https://github.com/aerogear/graphback/pull/1769)). 
+  This fixes issues below related to primary key in MongoDB:
+  - [#1731](https://github.com/aerogear/graphback/issues/1732), [#1626](https://github.com/aerogear/graphback/issues/1626) Ineffecient and wrong mapping of the `id` field
+  - [#1625](https://github.com/aerogear/graphback/issues/1625) Filters on "id: ID" field not working
+  
+  > NOTE: If you are migrating from [0.15](#0150) or previous versions of Graphback, you may be required to update relationship fields so that their values (previous stored as `String`) are of type [`ObjectID`](https://docs.mongodb.com/manual/reference/method/ObjectId/).
+ 
 # 0.15.1
 
 ### Bug Fixes
 
 * ensure field directives are mapped to the schema on relationship fields ([#1797](https://github.com/aerogear/graphback/pull/1797)) ([e8bf5c8](https://github.com/aerogear/graphback/commit/e8bf5c80633e51c3c9365a92f4f1ae6271ed6d2c))
-
 
 # 0.15.0